correction to hometask lesson 10
